+++ lib/AppConfig.php
@@ -559,7 +559,7 @@ class AppConfig {
             $this->logger->info("Set secret key", ["app" => $this->appName]);
         }
 
-        $this->appConfig->setValueString($this->appName, $this->_jwtSecret, $secret);
+        $this->appConfig->setValueString($this->appName, $this->_jwtSecret, $secret, sensitive: true);
     }
 
     /**
@@ -1156,7 +1156,7 @@ class AppConfig {
             return $turnOff === "true";
         }
 
-        return $this->getSystemValue($this->_verification) === "true";
+        return filter_var($this->getSystemValue($this->_verification), FILTER_VALIDATE_BOOLEAN);
     }
 
     /**